refactor whole ability pipeline

+++ lib/core/ecs/stores/status/slow_store.dart
@@ -5,13 +5,15 @@ class SlowDef {
   const SlowDef({required this.ticksLeft, required this.magnitude});
 
   final int ticksLeft;
-  final double magnitude;
+  /// Basis points (100 = 1%).
+  final int magnitude;
 }
 
 /// Active slow status (movement speed multiplier).
 class SlowStore extends SparseSet {
   final List<int> ticksLeft = <int>[];
-  final List<double> magnitude = <double>[];
+  /// Basis points (100 = 1%).
+  final List<int> magnitude = <int>[];
 
   void add(EntityId entity, SlowDef def) {
     final i = addEntity(entity);
@@ -22,7 +24,7 @@ class SlowStore extends SparseSet {
   @override
   void onDenseAdded(int denseIndex) {
     ticksLeft.add(0);
-    magnitude.add(0.0);
+    magnitude.add(0);
   }
 
   @override
@@ -33,4 +35,3 @@ class SlowStore extends SparseSet {
     magnitude.removeLast();
   }
 }
-
